Data Recording, Run Simulation Method, Performance Pertrubation Flag
Add average payoff time series recording within model.
Added the ability to record data to a csv.
Added the run_simulation method which runs the model for a given number of steps.
Added the ability to switch how the model perturbs agents beliefs.  Can perturb based on their performance or just with a perturbation probability parameter.